+++ pact-request-api/Pact/Core/Command/Types.hs
@@ -29,7 +29,7 @@
 -- Specifies types for commands in a consensus/DL setting.
 --
 
-module Pact.Core.Types.Command
+module Pact.Core.Command.Types
   ( Command(..),cmdPayload,cmdSigs,cmdHash
   , mkCommand
   , mkCommand'
@@ -89,11 +89,12 @@ import Pact.Core.DefPacts.Types
 import Pact.Core.Errors
 import Pact.Core.Gas.Types
 import Pact.Core.Guards
+import Pact.Core.Names
 import qualified Pact.Core.Hash as PactHash
 import Pact.Core.Persistence.Types
-import Pact.Core.Types.Orphans ()
+import Pact.Core.Command.Orphans ()
 import Pact.Core.PactValue (PactValue(..))
-import Pact.Core.Types.RPC
+import Pact.Core.Command.RPC
 import Pact.Core.IR.Eval.Runtime
 import Pact.Core.Verifiers
 
@@ -101,7 +102,7 @@ import Pact.JSON.Legacy.Value
 import qualified Pact.JSON.Encode as J
 
 
-import Pact.Core.Types.Crypto              as Base
+import Pact.Core.Command.Crypto              as Base
 
 -- | Command is the signed, hashed envelope of a Pact execution instruction or command.
 -- In 'Command ByteString', the 'ByteString' payload is hashed and signed; the ByteString
@@ -146,122 +147,11 @@ instance (NFData a,NFData m) => NFData (ProcessedCommand m a)
 
 type Ed25519KeyPairCaps = (Ed25519KeyPair ,[SigCapability])
 
--- CREATING AND SIGNING TRANSACTIONS
+-- These two types in legacy pact had the same definition and
+-- JSON encoding. Can they be unified?
+type UserCapability = CapToken QualifiedName PactValue
+type SigCapability = CapToken QualifiedName PactValue
 
-mkCommand
-  :: J.Encode c
-  => J.Encode m
-  => [(Ed25519KeyPair, [SigCapability])]
-  -> [Verifier ParsedVerifierProof]
-  -> m
-  -> Text
-  -> Maybe NetworkId
-  -> PactRPC c
-  -> IO (Command ByteString)
-mkCommand creds vers meta nonce nid rpc = mkCommand' creds encodedPayload
-  where
-    encodedPayload = J.encodeStrict $ toLegacyJsonViaEncode payload
-    payload = Payload rpc nonce meta (keyPairsToSigners creds) (vers <$ guard (not (null vers))) nid
-
-data WebAuthnPubKeyPrefixed
-  = WebAuthnPubKeyPrefixed
-  | WebAuthnPubKeyBare
-  deriving (Eq, Show, Generic)
-data DynKeyPair
-  = DynEd25519KeyPair Ed25519KeyPair
-  | DynWebAuthnKeyPair WebAuthnPubKeyPrefixed WebAuthnPublicKey WebauthnPrivateKey
-  deriving (Eq, Show, Generic)
-
-mkCommandWithDynKeys
-  :: J.Encode c
-  => J.Encode m
-  => [(DynKeyPair, [UserCapability])]
-  -> [Verifier ParsedVerifierProof]
-  -> m
-  -> Text
-  -> Maybe NetworkId
-  -> PactRPC c
-  -> IO (Command ByteString)
-mkCommandWithDynKeys creds vers meta nonce nid rpc = mkCommandWithDynKeys' creds encodedPayload
-  where
-    encodedPayload = J.encodeStrict $ toLegacyJsonViaEncode payload
-    payload = Payload rpc nonce meta (map credToSigner creds) (vers <$ guard (not (null vers))) nid
-    credToSigner cred =
-      case cred of
-        (DynEd25519KeyPair (pubEd25519, _), caps) ->
-          Signer
-            { _siScheme = Nothing
-            , _siPubKey = toB16Text (exportEd25519PubKey pubEd25519)
-            , _siAddress = Nothing
-            , _siCapList = caps
-            }
-        (DynWebAuthnKeyPair isPrefixed pubWebAuthn _, caps) ->
-          let
-            prefix = case isPrefixed of
-              WebAuthnPubKeyBare -> ""
-              WebAuthnPubKeyPrefixed -> webAuthnPrefix
-          in Signer
-            { _siScheme = Just WebAuthn
-            , _siPubKey = prefix <> toB16Text (exportWebAuthnPublicKey pubWebAuthn)
-            , _siAddress = Nothing
-            , _siCapList = caps
-            }
-
-keyPairToSigner :: Ed25519KeyPair -> [UserCapability] -> Signer
-keyPairToSigner cred caps = Signer scheme pub addr caps
-      where
-        scheme = Nothing
-        pub = toB16Text $ exportEd25519PubKey $ fst cred
-        addr = Nothing
-
-keyPairsToSigners :: [Ed25519KeyPairCaps] -> [Signer]
-keyPairsToSigners creds = map (uncurry keyPairToSigner) creds
-
-
-mkCommand' :: [(Ed25519KeyPair ,a)] -> ByteString -> IO (Command ByteString)
-mkCommand' creds env = do
-  let hsh = hash env    -- hash associated with a Command, aka a Command's Request Key
-      toUserSig (cred,_) = ED25519Sig $ signHash hsh cred
-  let sigs = toUserSig <$> creds
-  return $ Command env sigs hsh
-
--- | A utility function used for testing.
--- It generalizes `mkCommand` by taking a `DynKeyPair`, which could contain mock
--- WebAuthn keys. If WebAuthn keys are encountered, this function does mock WebAuthn
--- signature generation when constructing the `Command`.
-mkCommandWithDynKeys' :: [(DynKeyPair, a)] -> ByteString -> IO (Command ByteString)
-mkCommandWithDynKeys' creds env = do
-  let hsh = hash env    -- hash associated with a Command, aka a Command's Request Key
-  sigs <- traverse (toUserSig hsh) creds
-  return $ Command env sigs hsh
-  where
-    toUserSig :: PactHash -> (DynKeyPair, a) -> IO UserSig
-    toUserSig hsh = \case
-      (DynEd25519KeyPair (pub, priv), _) ->
-        pure $ ED25519Sig $ signHash hsh (pub, priv)
-      (DynWebAuthnKeyPair _ pubWebAuthn privWebAuthn, _) -> do
-        signResult <- runExceptT $ signWebauthn pubWebAuthn privWebAuthn "" (toUntypedHash hsh)
-        case signResult of
-          Left e -> error $ "Failed to sign with mock WebAuthn keypair: " ++ e
-          Right sig -> return $ WebAuthnSig sig
-
-mkUnsignedCommand
-  :: J.Encode m
-  => J.Encode c
-  => [Signer]
-  -> [Verifier ParsedVerifierProof]
-  -> m
-  -> Text
-  -> Maybe NetworkId
-  -> PactRPC c
-  -> IO (Command ByteString)
-mkUnsignedCommand signers vers meta nonce nid rpc = mkCommand' [] encodedPayload
-  where encodedPayload = J.encodeStrict payload
-        payload = Payload rpc nonce meta signers (vers <$ guard (not (null vers))) nid
-
-signHash :: TypedHash h -> Ed25519KeyPair -> Text
-signHash hsh (pub,priv) =
-  toB16Text $ exportEd25519Signature $ signEd25519 pub priv (toUntypedHash hsh)
 
 -- VALIDATING TRANSACTIONS
 
@@ -283,12 +173,12 @@ verifyCommand orig@Command{..} =
 
     verifiedHash = verifyHash _cmdHash _cmdPayload
 
-hasInvalidSigs :: PactHash -> [UserSig] -> [Signer] -> Maybe String
+hasInvalidSigs :: PactHash.Hash -> [UserSig] -> [Signer] -> Maybe String
 hasInvalidSigs hsh sigs signers
   | not (length sigs == length signers)  = Just "Number of sig(s) does not match number of signer(s)"
   | otherwise                            = verifyUserSigs hsh (zip sigs signers)
 
-verifyUserSigs :: PactHash -> [(UserSig, Signer)] -> Maybe String
+verifyUserSigs :: PactHash.Hash -> [(UserSig, Signer)] -> Maybe String
 verifyUserSigs hsh sigsAndSigners
   | null failedSigs = Nothing
   | otherwise = formatIssues
@@ -299,7 +189,7 @@ verifyUserSigs hsh sigsAndSigners
   failedSigs = concatMap getFailedVerify sigsAndSigners
   formatIssues = Just $ "Invalid sig(s) found: " ++ show (J.encode . J.Object $ failedSigs)
 
-verifyUserSig :: PactHash -> UserSig -> Signer -> Either String ()
+verifyUserSig :: PactHash.Hash -> UserSig -> Signer -> Either String ()
 verifyUserSig msg sig Signer{..} = do
   case (sig, scheme) of
     (ED25519Sig edSig, ED25519) -> do
@@ -332,6 +222,26 @@ verifyUserSig msg sig Signer{..} = do
         ]
   where scheme = fromMaybe defPPKScheme _siScheme
 
+
+instance J.Encode Signer where
+  build o = J.object
+    [ "addr" J..?= _siAddress o
+    , "scheme" J..?= _siScheme o
+    , "pubKey" J..= _siPubKey o
+    , "clist" J..??= J.Array (_siCapList o)
+    ]
+
+instance FromJSON Signer where
+  parseJSON = withObject "Signer" $ \o -> Signer
+    <$> o .:? "scheme"
+    <*> o .: "pubKey"
+    <*> o .:? "addr"
+    <*> (listMay <$> (o .:? "clist"))
+    where
+      listMay = fromMaybe []
+
+instance Arbitrary Signer where
+  arbitrary = Signer <$> arbitrary <*> arbitrary <*> arbitrary <*> scale (min 5) arbitrary
 
 -- | Payload combines a 'PactRPC' with a nonce and platform-specific metadata.
 data Payload m c = Payload
